added console interface executeCode functionallity

+++ project0/Store/Customer.cs
@@ -2,11 +2,11 @@ namespace Store
 {
     public class Customer
     {
-        private static ulong customerTally;
-        private ulong customerId;
+        private static uint customerTally;
+        private uint customerId;
         private string customerName;
         public string CustomerName { get => customerName; }
-        public ulong CustomerId { get => customerId; }
+        public uint CustomerId { get => customerId; }
         public Customer(string name) {
             this.customerId = customerTally++;
             this.customerName = name;
